refactor to use new audio-gen node execution architecture

+++ src/graph.rs
@@ -4,6 +4,6 @@ use crate::JSNode;
 
 #[derive(Deserialize, Serialize)]
 pub struct SourceGraph {
-    pub root_id: i32,
+    pub root_id: usize,
     pub nodes: Vec<JSNode>,
 }
